requirements.txt
jupyterlab==4.0.7
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pydantic==2.4.2
pandas==2.1.1
tenacity==8.2.3
python-dotenv==1.0.0
tqdm==4.66.1
pytz==2023.3.post1
datasets==2.15.0
scikit_learn==1.3.2
openai==1.3.2
accelerate==0.21.0
peft==0.6.2
bitsandbytes==0.41.2.post2
transformers==4.35.2
trl==0.4.7
torch==2.1.0